requirements.txt
fastapi
uvicorn
tensorflow
python-dotenv
python-multipart
python-jose
sqlalchemy
psycopg2
bcrypt
passlib
pandas
joblib
Pillow
scikit-learn==1.3.2